cmd: PPC: support tolerating heterogeneous core IDs
Problem: In a system with a specific number of cores per socket
the host gives each core a number. So far PPC would generate
performance profile only after verifying that the nodes pointed to by
the specified node pool are all of the same hardware and topology. That
is because the performance profile will apply the same configuration on all
these nodes and it is most important that they have same structure, CPUs
distribution across NUMAs and CPUs availability. This by default
includes having same core IDs for each CPU in the same NUMA cells for
all the compute nodes. for instance:

Worker-0 has NUMA-0 on which the CPUs siblings [2,66] is coupled in
one core numbered 18`.
All other workers should have this info true for them, otherwise the
tool would fail.

It was observed (for example on Intel Xeon GoldGold 6438N with 0-127
online CPUs distributed across 2 sockets, 32 cores per socket and 2 threads per core)
that core numbering can have different schemes even with a system from the same vendor,
which causes the tool to fail to generate a profile.

Suggested solution:
With further investigation, The numbering the pattern depends on the settings
of the hardware, the software and the firmware (BIOS).While core IDs may vary
nodes can still be considered having same NUMA topology taking into account
that core scope is on the single NUMA. However, core IDs can be important in
optimizing the system's performance and managing isolation of tasks, meaning if the
performance of worker-0 is not comparable to that of worker-1 before
having performance-profile applied on both, having an improved
performance on both after applying PP is not guaranteed.

In this PR, we loosen the hard requirment of having same core numbering
on same NUMA cells on different systems into a warning will be logged
out as well as a comment on the generated CSV file. So as long as the NUMA cells
have same logical processors' count and IDs and same threads' number,
core ID equality is treated as best effort. That is because when scheduling workloads,
we care about the logical processors ids and their location on the NUMAs.

Disclaimer: We support this option to unblock business matters and is recommended
to use the generated PP with cautiono.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>